+++ config/initializers/widget_types.rb
@@ -1,4 +1,0 @@
-Pageflow.configure do |config|
-  config.widget_types.register(Pageflow::BuiltInWidgetType.navigation, default: true)
-  config.widget_types.register(Pageflow::BuiltInWidgetType.mobile_navigation, default: true)
-end