+++ spec/generators/pageflow/initializer/initializer_generator_spec.rb
@@ -15,14 +15,24 @@ module Pageflow
         expect(initializer).to exist
       end
 
-      it 'registers classic player controls' do
+      it 'registers the built-in classic player controls' do
         expect(initializer)
           .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.classic_player_controls')
       end
 
-      it 'registers slim player controls' do
+      it 'registers the built-in slim player controls' do
         expect(initializer)
           .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.slim_player_controls')
+      end
+
+      it 'registers the built-in navigation widget type' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.navigation')
+      end
+
+      it 'registers the built-in mobile_navigation widget type' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.mobile_navigation')
       end
     end
   end