+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 
 #### Breaking Changes
 
-- Move player controls initialization to the generated Pageflow initializer. Add these lines to `config/initializers/pageflow.rb` in your host application:
+- The built-in widget types must now be registered in the host application. To keep existing functionality from previous Pageflow versions, add these lines to `config/initializers/pageflow.rb` in your host application:
 
 ```
-# Configure the classic or slim player controls.
+# Register the built-in widget types.
+# You can remove these or add different versions with the same name.
+config.widget_types.register(Pageflow::BuiltInWidgetType.navigation, default: true)
+config.widget_types.register(Pageflow::BuiltInWidgetType.mobile_navigation, default: true)
 config.widget_types.register(Pageflow::BuiltInWidgetType.classic_player_controls, default: true)
 config.widget_types.register(Pageflow::BuiltInWidgetType.slim_player_controls)
 ```